requirements.txt
flask
werkzeug
python-docx
langchain
faiss-cpu
sentence-transformers
unstructured
pytesseract
pdf2image
Pillow
requests
beautifulsoup4
openpyxl